Added a new S3Client class and companion object to wrap the AmazonS3 client object. Also split the S3PartitionReader out into its own file.
