+++ src/main/scala/com/github/datasource/s3/S3DataSource.scala
@@ -152,51 +152,3 @@ class S3PartitionReaderFactory(pushdown: Pushdown,
   }
 }
 
-/** PartitionReader of S3Partitions
- *
- * @param pushdown object handling filter, project and aggregate pushdown
- * @param options the options including "path"
- */
-class S3PartitionReader(pushdown: Pushdown,
-                        options: util.Map[String, String],
-                        partition: S3Partition)
-  extends PartitionReader[InternalRow] {
-
-  private val logger = LoggerFactory.getLogger(getClass)
-
-  /* We setup a rowIterator and then read/parse
-   * each row as it is asked for.
-   */
-  private var store: S3Store = S3StoreFactory.getS3Store(pushdown, options)
-  private var rowIterator: Iterator[InternalRow] = store.getRowIter(partition)
-  private var rows: Long = if (options.containsKey("EnableProgress")) {
-    store.getNumRows
-  } else {
-    Long.MaxValue
-  }
-  var index: Long = 0
-  def next: Boolean = {
-    rowIterator.hasNext
-  }
-  def get: InternalRow = {
-    val row = rowIterator.next
-    if (((index % 500000) == 0) ||
-        (!next)) {
-      val pct: String = {
-        if (rows == Long.MaxValue ||
-            rows == 0) {
-          ""
-        } else {
-          val percentage = ((index * 100.toDouble) / rows.toDouble)
-          s"/${rows} " + "%.2f".format(percentage).toDouble.toString + " %"
-        }
-      }
-      logger.info(s"get: partition: ${partition.index} ${partition.bucket} " +
-                  s"${partition.key} index: ${index}${pct}")
-    }
-    index = index + 1
-    row
-  }
-
-  def close(): Unit = Unit
-}
